cmd handles errors
